progress: remove duplicate mod progress::progress
